add HandleAndCache to InstanceEntity
