+++ crates/wasmi/src/instance/mod.rs
@@ -1,17 +1,16 @@
 pub(crate) use self::builder::InstanceEntityBuilder;
 use self::handle::AnyHandle;
 pub use self::{
+    cache::HandleAndCache,
     exports::{Export, ExportsIter, Extern, ExternType},
     layout::{DataAddr, ElemAddr, FuncAddr, GlobalAddr, InstanceLayout, MemoryAddr, TableAddr},
 };
 use crate::{
     AsContext,
     AsContextMut,
-    DataSegmentEntity,
     ElementSegment,
     Error,
     Func,
-    FuncEntity,
     Global,
     Memory,
     Module,
@@ -21,21 +20,15 @@ use crate::{
     WasmParams,
     WasmResults,
     collections::Map,
-    core::{
-        CoreElementSegment as ElementSegmentEntity,
-        CoreGlobal as GlobalEntity,
-        CoreMemory as MemoryEntity,
-        CoreTable as TableEntity,
-    },
     engine::DedupFuncType,
     func::FuncError,
     memory::DataSegment,
-    store::{StoreInner, Stored},
+    store::Stored,
 };
 use alloc::{boxed::Box, sync::Arc};
-use core::ptr::NonNull;
 
 mod builder;
+mod cache;
 mod exports;
 mod handle;
 mod layout;
@@ -46,119 +39,12 @@ mod tests;
 /// A module instance entity.
 #[derive(Debug)]
 pub struct InstanceEntity {
-    handles: Box<[AnyHandle]>,
+    handles: Box<[HandleAndCache]>,
     func_types: Arc<[DedupFuncType]>,
     exports: Map<Box<str>, Extern>,
     layout: InstanceLayout,
     initialized: bool,
 }
-
-/// An [`AnyHandle`] and its optional cached entity for hot reloading.
-#[derive(Debug)]
-pub struct HandleAndCache {
-    /// The cached entity for hot reloading.
-    cache: Option<NonNull<AnyEntity>>,
-    /// The entity handle.
-    handle: AnyHandle,
-}
-
-impl HandleAndCache {
-    /// Creates a new [`HandleAndCache`] from the given `handle`.
-    pub fn new(handle: AnyHandle) -> Self {
-        Self {
-            cache: None,
-            handle,
-        }
-    }
-}
-
-macro_rules! impl_handle_and_cache {
-    (
-        $(
-            pub unsafe fn $name:ident<'a>(&mut self, store: &'a mut StoreInner) -> Option<&'a mut $ty:ty> = {
-                resolve: $resolve:expr,
-                cast: $cast:expr,
-                load: $load:ident,
-            }
-        )*
-    ) => {
-        $(
-            #[doc = concat!("Returns the [`", stringify!($ty), "`] for `self` if any.")]
-            #[doc = ""]
-            #[doc = "# Safety"]
-            #[doc = ""]
-            #[doc = "It is the caller's responsibility to use this only if the"]
-            #[doc = concat!("[`HandleAndCache`] is associated to a [`", stringify!($ty), "`].")]
-            #[inline]
-            pub unsafe fn $name<'a>(&mut self, store: &'a mut StoreInner) -> Option<&'a mut $ty> {
-                if let Some(cache) = &mut self.cache {
-                    // Case: cache already exists and can be re-used.
-                    let entity = unsafe { cache.cast::<$ty>().as_mut() };
-                    return Some(entity)
-                }
-                // Case: cache is vacant and the entity needs to be loaded from the `store`.
-                unsafe { Self::$load(self, store) }
-            }
-
-            #[doc = concat!("Loads the [`", stringify!($ty), "`] for `self` from `store` and caches it.")]
-            #[doc = ""]
-            #[doc = "# Safety"]
-            #[doc = ""]
-            #[doc = "It is the caller's responsibility to use this only if the"]
-            #[doc = concat!("[`HandleAndCache`] is associated to a [`", stringify!($ty), "`].")]
-            #[cold]
-            unsafe fn $load<'a>(&mut self, store: &'a mut StoreInner) -> Option<&'a mut $ty> {
-                let handle = unsafe { $cast(self.handle) };
-                let entity = $resolve(store, &handle);
-                self.cache = Some(NonNull::from(&mut *entity).cast::<AnyEntity>());
-                Some(entity)
-            }
-        )*
-    };
-}
-impl HandleAndCache {
-    impl_handle_and_cache! {
-        pub unsafe fn get_memory<'a>(&mut self, store: &'a mut StoreInner) -> Option<&'a mut MemoryEntity> = {
-            resolve: StoreInner::resolve_memory_mut,
-            cast: AnyHandle::cast_memory,
-            load: load_memory,
-        }
-
-        pub unsafe fn get_global<'a>(&mut self, store: &'a mut StoreInner) -> Option<&'a mut GlobalEntity> = {
-            resolve: StoreInner::resolve_global_mut,
-            cast: AnyHandle::cast_global,
-            load: load_global,
-        }
-
-        pub unsafe fn get_table<'a>(&mut self, store: &'a mut StoreInner) -> Option<&'a mut TableEntity> = {
-            resolve: StoreInner::resolve_table_mut,
-            cast: AnyHandle::cast_table,
-            load: load_table,
-        }
-
-        pub unsafe fn get_func<'a>(&mut self, store: &'a mut StoreInner) -> Option<&'a mut FuncEntity> = {
-            resolve: StoreInner::resolve_func_mut,
-            cast: AnyHandle::cast_func,
-            load: load_func,
-        }
-
-        pub unsafe fn get_elem<'a>(&mut self, store: &'a mut StoreInner) -> Option<&'a mut ElementSegmentEntity> = {
-            resolve: StoreInner::resolve_element_mut,
-            cast: AnyHandle::cast_elem,
-            load: load_elem,
-        }
-
-        pub unsafe fn get_data<'a>(&mut self, store: &'a mut StoreInner) -> Option<&'a mut DataSegmentEntity> = {
-            resolve: StoreInner::resolve_data_mut,
-            cast: AnyHandle::cast_data,
-            load: load_data,
-        }
-    }
-}
-
-/// Represents any entity kind and used for type pruning in [`HandleAndCache`].
-#[derive(Debug)]
-pub struct AnyEntity;
 
 impl InstanceEntity {
     /// Creates an uninitialized [`InstanceEntity`].
@@ -190,37 +76,37 @@ impl InstanceEntity {
     /// Returns the [`Memory`] at the `addr` if any.
     pub fn get_memory(&self, addr: MemoryAddr) -> Option<Memory> {
         let handle = self.handles.get(u32::from(addr) as usize)?;
-        Some(unsafe { handle.cast_memory() })
+        Some(unsafe { handle.handle.cast_memory() })
     }
 
     /// Returns the [`Table`] at the `addr` if any.
     pub fn get_table(&self, addr: TableAddr) -> Option<Table> {
         let handle = self.handles.get(u32::from(addr) as usize)?;
-        Some(unsafe { handle.cast_table() })
+        Some(unsafe { handle.handle.cast_table() })
     }
 
     /// Returns the [`Global`] at the `addr` if any.
     pub fn get_global(&self, addr: GlobalAddr) -> Option<Global> {
         let handle = self.handles.get(u32::from(addr) as usize)?;
-        Some(unsafe { handle.cast_global() })
+        Some(unsafe { handle.handle.cast_global() })
     }
 
     /// Returns the [`Func`] at the `addr` if any.
     pub fn get_func(&self, addr: FuncAddr) -> Option<Func> {
         let handle = self.handles.get(u32::from(addr) as usize)?;
-        Some(unsafe { handle.cast_func() })
+        Some(unsafe { handle.handle.cast_func() })
     }
 
     /// Returns the [`DataSegment`] at the `addr` if any.
     pub fn get_data_segment(&self, addr: DataAddr) -> Option<DataSegment> {
         let handle = self.handles.get(u32::from(addr) as usize)?;
-        Some(unsafe { handle.cast_data() })
+        Some(unsafe { handle.handle.cast_data() })
     }
 
     /// Returns the [`ElementSegment`] at the `addr` if any.
     pub fn get_element_segment(&self, addr: ElemAddr) -> Option<ElementSegment> {
         let handle = self.handles.get(u32::from(addr) as usize)?;
-        Some(unsafe { handle.cast_elem() })
+        Some(unsafe { handle.handle.cast_elem() })
     }
 
     /// Returns the signature at the `index` if any.
